- Updated the directory structure, pulled out Tabs navigation so it would be less confusing as we scale where tabs screens are vs other screens
