+++ App.tsx
@@ -1,6 +1,6 @@
 import { AuthProvider } from "./src/auth0";
 import LoginScreen from "./src/screens/LoginScreen";
-import MapsScreen from "./src/screens/MapsScreen";
+import MapTabDrawerScreen from "./src/TabsNavigation/MapTabDrawerScreen";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
@@ -14,7 +14,7 @@ export default function App() {
         <NavigationContainer>
           <Stack.Navigator screenOptions={{ headerShown: false }}>
             <Stack.Screen name="Login" component={LoginScreen} />
-            <Stack.Screen name="MapsScreen" component={MapsScreen} />
+            <Stack.Screen name="MapScreen" component={MapTabDrawerScreen} />
           </Stack.Navigator>
         </NavigationContainer>
       </SafeAreaProvider>
