+++ src/TabsNavigation/types.ts
@@ -1,0 +1,10 @@
+export type RootStackParamList = {
+  MapScreen: undefined,
+  LoginScreen : undefined,
+  MapTabDrawerScreen : undefined,
+  OfflineMapScreen : undefined,
+  RoutesTabScreen : undefined,
+  SettingsScreen: undefined,
+  UserProfileScreen: undefined,
+  UserProfileTabScreen : undefined,
+}
